Move warn_if_run_as_root() to utils
While it's reasonable to put this function in req_install.py, putting it
there unfortunately means importing a bunch of heavy network/index code
as well which isn't acceptable.
